Extract into new files
